+++ microimpute/comparisons/__init__.py
@@ -6,9 +6,6 @@ This module contains utilities for comparing different imputation methods.
 # Import automated imputation utilities
 from .autoimpute import autoimpute
 
-# Import data handling functions
-from .data import prepare_scf_data, preprocess_data, scf_url
-
 # Import imputation utilities
 from .imputations import get_imputations
