+++ tests/test_models/test_imputers.py
@@ -12,7 +12,7 @@ import pandas as pd
 import pytest
 from sklearn.datasets import load_diabetes
 
-from microimpute.comparisons.data import preprocess_data
+from microimpute.utils.data import preprocess_data
 from microimpute.config import QUANTILES
 from microimpute.models import *
